Updating config.md
Adds section for Testbench Macros which are automatically generated.
Moves a few of the existing macros into this section, with additional information for how they can be modified in the config.
Adds all of the defines I could find in genchecks as well for consistency.
Changes the examples given for the isa option and notes that X and Z extensions are not supported until we have proper parsing instead of just checking if (e.g.) 'c' appears in the string.