Better hpm checking
Introduces csrc_upcnt_check.
Similar to inc, but instead of checking if the value is greater than or equal to the value written/read it specifically disallows writes to the register being tested.
In order to support all of the hpm CSRs, the corresponding hpmevent CSR must be written and non-zero prior to checking the hpmcounter CSR.
Note that this may require liveness guarantees depending on the hpmevents supported.
Currently lacking documentation.