Updating picorv32 for new csrc tests
Specifically, mcycle and minstret are now tested under inc and upcnt.
Adds test to ensure writes are disallowed to umode CSRs of mcycle and minstret, since they should be read-only.
Note that technically at present it is not possible to write to these registers even in mmode, and they are not available at their mmode address.  But it's something.

Also removes unused PICORV32_CSR_RESTRICT flag.