Working B instructions
In theory it's still backwards compatible with rv32, but not tested.
CSRs, memory, and probably other things are still 32 bit.  This is very much *not* a fully implemented rv64 core.  It *is* however enough to test the rv64b instructions.
This also strips out the NERV_FAULT and NERV_FAIRNESS stuff from the wrapper.