Proto C extension
Extensions can register validity checks and instruction-aware assignments.
Used with the C extension to inject padding checks to all instructions, but only if the C extension is loaded in Python, and assigning the padding based on the ilen of the instruction.
Drop `ilen` from the `to_verilog()` method, and instead make it an Instruction variable.  Necessary for the padding detection, but also we already have `RISCV_FORMAL_ILEN` for the RVFI field size, and what we want to know is the instruction's actual width.
Also make opcode_width and default pc increment instruction variables.  The former being public and the latter being private.