Moving sail things out of the base model
Replace base `inst_args` with `_insn_part_dict.keys()`.
Drop `wrap_next_pc` from base, instead make `_v_spec_value()` have an optional return, so that when it returns `None` the value key doesn't get added to the spec map.
Also update the sail makefile so that it outputs the current value of `INSN` to a file called `insn` if it doesn't match.  This lets us use that file as a pre-requisite instead of making the `INSN` specific files `PHONY` (and therefore we are able to edit the generated file, e.g. to add an `assume (rvfi.trap == 0);`).