Add memory fault handling
Accomplished via registered hw traps in the InstructionChecker class.
Fault handlers are able to use the observers dict to generate only the necessary assertions, or raise an error if something is missing.  But currently the mem fault is just a fixed string.
Update the notsail Makefile to support model flags.
Add nerv_fault target to rvf_complete.sby.