Add misa faults
Can be either strong (if misa bit(s) are unset it *must* trap) or weak (may or may not trap).
Add more `InstructionChecker` registrations, allows registering custom checks, raw verilog, and speculators.
Checks get pre-pended to `defined_checks`.
Raw verilog is injected at the end of `_v_instantiation()`.
Speculators register a `SpeculativeObserver` (providing the name, bitrange and default value for the speculative signal) and a callable which takes a given instruction and optionally returns a value to be assigned to the speculative signal when that instruction is valid.

I tested this by adding a MISA test to NERV for some of the Zbb instructions where if the "B" bit is unset then it is instead an `illinsn` and it appears to work as expected.  The weak misa check passes all instructions, while the strong misa check only passes the instructions with the MISA test.