Adding csrw check
But only for rvfi enabled CSRs.
Move around some signals so that they are more useful (e.g. we can `assume(csr_access)` without also assuming that the privilege level is valid, and can therefore assert that if the privilege level is *not* valid then the access must be illegal).
CSRW check is independent of the CSR behavior, so we place it separately (and optionally).