csr_ill_check from Python
Uses `GenericGroupChecker` class with a list of valid CSRs.
Read/write access to any other CSR address should cause a trap (including access to *h registers when not rv32).
Illegal access to valid CSRs (writing a read-only CSR or too-low privilege) is handled by the RW check, so we don't need it here.