NERV: Playing with MISA controlled instructions
Some of the Zbb instructions (CLZ, CTZ, CPOP, and SEXT) are now executed only if the B flag in MISA is set.  If the flag is not set then an illegal instruction is raised instead.
This is mostly just for testing that instruction checks are still valid when instruction availability can be conditional on MISA.