Progress on CSRs in instruction checks
Add `has_macro_define` (i.e. there is a corresponding `RISCV_FORMAL_CSR_*`), which is true for all built in CSRs and false for all custom/user provided CSRs (which instead need to be in the `RISCV_FORMAL_CUSTOM_CSR_*` defines).
Fix missing `csr_none`.
Add `CsrInstruction` class, which has a list of the CSRs accessed in order to add them to the inputs used when selecting IO.  Haven't yet figured out how to get outputs with speculative values, but reading CSRs works (provided the rmask is always fully set).