NERV: Add custom extension for testing
`custom` and `custom_ro` are now `dummy` and `dummy_ro`, and are defined by the Xnerv extension.
Xnerv also provides a single instruction, `dummy_test` which uses a custom instruction format in the custom-0 opcode space to add a sign extended 21bit immediate (where the 4 LSb are 0 and the 17 MSb are encoded by the instruction) to the value in `dummy` and write the result to `rd`.
This extension is only enabled in NERV if `NERV_EXT` is defined.

Also comment out the Zicntr and Zihpm csrs since we now check `isa` for which csrs are available and they are not available in the current `isa` config.