genchecks: CSR behaviors
Or at least `const`.
Make the channel iteration and `gen_check` awaiting as it's own function, and split the different checks back into their own for loops, now (mostly) with a single line.
CSR check generation has some extra logic for behavioral testing (i.e. `csrc` checks), including (temporarily) clearing the default behavior to make sure we get the RW test.
Add debug print on successful sby file generation.
`GenInsnCheck` can now have an expected depth of 2 (for `csrc` checks), which assigns a value to `@start@`.
Also I noticed that the `_mask=` behavior on csrc checks is broken, so add a TODO for that (and simplify the regex for finding tests slightly by using single quotes for the string so we don't have to escape the double quotes in the regex).