Re-enable other csrc checks
`behavior.short_name` -> `Behavior.NAME` (i.e. it's now a constant class var, set with `__init_subclass__`.
Behavior types are registered with the `CsrSpec`, using the `NAME` as the key.
`genchecks` then uses the `CsrSpec.get_behavior()` to lookup behaviors for generation.
Fix default behavior not being reset if it was `None` (meaning the csr could get left with the last test behavior, probably not a huge issue but better to be safe).
Fix duplicate names for csrc checks (it was previously doing both the csrw and csrc in the same file so had to filter on both, but that is no longer the case).